+++ integration-tests/global-teardown/invalid_teardown.js
@@ -1,0 +1,1 @@
+console.log('there is no exported function');