+++ integration-tests/typescript-coverage/covered.ts
@@ -2,4 +2,4 @@
 
 export = function sum(a: number, b: number): number {
   return a + b;
-}
+};