+++ packages/expect/src/spy_matchers.js
@@ -24,6 +24,7 @@ import {
 } from 'jest-matcher-utils';
 import {equals} from './jasmine_utils';
 import {iterableEquality, partition} from './utils';
+import diff from 'jest-diff';
 
 const createToBeCalledMatcher = matcherName => (received, expected) => {
   ensureNoExpected(expected, matcherName);
@@ -117,6 +118,46 @@ const createLastCalledWithMatcher = matcherName => (
 
 const spyMatchers: MatchersObject = {
   lastCalledWith: createLastCalledWithMatcher('.lastCalledWith'),
+  nthCalledWith(received: any, nth: number, ...expected: any) {
+    const matcherName = '.nthCalledWith';
+    ensureMock(received, matcherName);
+
+    const receivedIsSpy = isSpy(received);
+    const type = receivedIsSpy ? 'spy' : 'mock function';
+
+    if (typeof nth !== 'number' || parseInt(nth, 10) !== nth || nth < 1) {
+      const message = () =>
+        `nth value ${printReceived(
+          nth,
+        )} must be a positive integer greater than ${printExpected(0)}`;
+      const pass = false;
+      return {message, pass};
+    }
+
+    const receivedName = receivedIsSpy ? 'spy' : received.getMockName();
+    const calls = receivedIsSpy
+      ? received.calls.all().map(x => x.args)
+      : received.mock.calls;
+    const pass = equals(calls[nth - 1], expected, [iterableEquality]);
+
+    const message = pass
+      ? () =>
+          matcherHint('.not' + matcherName, receivedName) +
+          '\n\n' +
+          `Expected ${type} ${nthToString(
+            nth,
+          )} call to not have been called with:\n` +
+          `  ${printExpected(expected)}`
+      : () =>
+          matcherHint(matcherName, receivedName) +
+          '\n\n' +
+          `Expected ${type} ${nthToString(
+            nth,
+          )} call to have been called with:\n` +
+          formatMismatchedCalls(calls, expected, LAST_CALL_PRINT_LIMIT);
+
+    return {message, pass};
+  },
   toBeCalled: createToBeCalledMatcher('.toBeCalled'),
   toBeCalledWith: createToBeCalledWithMatcher('.toBeCalledWith'),
   toHaveBeenCalled: createToBeCalledMatcher('.toHaveBeenCalled'),
@@ -228,9 +269,12 @@ const formatMismatchedArgs = (expected, received) => {
   const printedArgs = [];
   for (let i = 0; i < length; i++) {
     if (!equals(expected[i], received[i], [iterableEquality])) {
+      const diffString = diff(expected[i], received[i]);
       printedArgs.push(
-        `  ${printExpected(expected[i])} as argument ${i + 1}, ` +
-          `but it was called with ${printReceived(received[i])}.`,
+        `  ${printExpected(expected[i])}\n` +
+          `as argument ${i + 1}, but it was called with\n` +
+          `  ${printReceived(received[i])}.` +
+          (diffString ? `\n\nDifference:\n\n${diffString}` : ''),
       );
     } else if (i >= expected.length) {
       printedArgs.push(
@@ -241,6 +285,18 @@ const formatMismatchedArgs = (expected, received) => {
   }
 
   return printedArgs.join('\n');
+};
+
+const nthToString = (nth: number) => {
+  switch (nth) {
+    case 1:
+      return 'first';
+    case 2:
+      return 'second';
+    case 3:
+      return 'third';
+  }
+  return `${nth}th`;
 };
 
 export default spyMatchers;