+++ packages/jest-cli/src/__tests__/test_scheduler.test.js
@@ -12,6 +12,20 @@ import TestScheduler from '../test_scheduler';
 import SummaryReporter from '../reporters/summary_reporter';
 
 jest.mock('../reporters/default_reporter');
+const mockSerialRunner = {
+  isSerial: true,
+  runTests: jest.fn(),
+};
+jest.mock('jest-runner-serial', () => jest.fn(() => mockSerialRunner), {
+  virtual: true,
+});
+
+const mockParallelRunner = {
+  runTests: jest.fn(),
+};
+jest.mock('jest-runner-parallel', () => jest.fn(() => mockParallelRunner), {
+  virtual: true,
+});
 
 test('.addReporter() .removeReporter()', () => {
   const scheduler = new TestScheduler({}, {});
@@ -20,4 +34,46 @@ test('.addReporter() .removeReporter()', () => {
   expect(scheduler._dispatcher._reporters).toContain(reporter);
   scheduler.removeReporter(SummaryReporter);
   expect(scheduler._dispatcher._reporters).not.toContain(reporter);
+});
+
+test('schedule tests run in parallel per default', async () => {
+  const scheduler = new TestScheduler({}, {});
+  const test = {
+    context: {
+      config: {
+        runner: 'jest-runner-parallel',
+      },
+      hasteFS: {
+        matchFiles: jest.fn(() => []),
+      },
+    },
+    path: './test/path.js',
+  };
+  const tests = [test, test];
+
+  await scheduler.scheduleTests(tests, {isInterrupted: jest.fn()});
+
+  expect(mockParallelRunner.runTests).toHaveBeenCalled();
+  expect(mockParallelRunner.runTests.mock.calls[0][5].serial).toBeFalsy();
+});
+
+test('schedule tests run in serial if the runner flags them', async () => {
+  const scheduler = new TestScheduler({}, {});
+  const test = {
+    context: {
+      config: {
+        runner: 'jest-runner-serial',
+      },
+      hasteFS: {
+        matchFiles: jest.fn(() => []),
+      },
+    },
+    path: './test/path.js',
+  };
+
+  const tests = [test, test];
+  await scheduler.scheduleTests(tests, {isInterrupted: jest.fn()});
+
+  expect(mockSerialRunner.runTests).toHaveBeenCalled();
+  expect(mockSerialRunner.runTests.mock.calls[0][5].serial).toBeTruthy();
 });