+++ packages/jest-cli/src/watch.js
@@ -82,17 +82,19 @@ export default function watch(
   });
 
   const updateConfigAndRun = ({
+    mode,
     testNamePattern,
     testPathPattern,
     updateSnapshot,
   }: {
+    mode?: 'watch' | 'watchAll',
     testNamePattern?: string,
     testPathPattern?: string,
     updateSnapshot?: SnapshotUpdateState,
   } = {}) => {
     const previousUpdateSnapshot = globalConfig.updateSnapshot;
     globalConfig = updateGlobalConfig(globalConfig, {
-      mode: 'watch',
+      mode,
       testNamePattern:
         testNamePattern !== undefined
           ? testNamePattern
@@ -152,6 +154,18 @@ export default function watch(
   let shouldDisplayWatchUsage = true;
   let isWatchUsageDisplayed = false;
 
+  const emitFileChange = () => {
+    if (hooks.isUsed('fileChange')) {
+      const projects = searchSources.map(({context, searchSource}) => ({
+        config: context.config,
+        testPaths: searchSource.findMatchingTests('').tests.map(t => t.path),
+      }));
+      hooks.getEmitter().fileChange({projects});
+    }
+  };
+
+  emitFileChange();
+
   hasteMapInstances.forEach((hasteMapInstance, index) => {
     hasteMapInstance.on('change', ({eventsQueue, hasteFS, moduleMap}) => {
       const validPaths = eventsQueue.filter(({filePath}) => {
@@ -174,6 +188,7 @@ export default function watch(
           context,
           searchSource: new SearchSource(context),
         };
+        emitFileChange();
         startRun(globalConfig);
       }
     });
@@ -312,6 +327,7 @@ export default function watch(
         break;
       case KEYS.C:
         updateConfigAndRun({
+          mode: 'watch',
           testNamePattern: '',
           testPathPattern: '',
         });