+++ packages/jest-jasmine2/src/jasmine/Suite.js
@@ -210,7 +210,7 @@ function convertDescriptorToString(descriptor) {
 }
 
 function isAfterAll(children) {
-  return children && children[0].result.status;
+  return children && children[0] && children[0].result.status;
 }
 
 function isFailure(args) {