+++ CHANGELOG.md
@@ -2,13 +2,107 @@
 
 ### Features
 
-* `[jest-runtime]` Provide `require.main` property set to module with test suite
-  ([#5618](https://github.com/facebook/jest/pull/5618))
+* `[expect]` Improve output format for mismatchedArgs in mock/spy calls.
+  ([#5846](https://github.com/facebook/jest/pull/5846))
+* `[jest-cli]` Add support for using `--coverage` in combination with watch
+  mode, `--onlyChanged`, `--findRelatedTests` and more
+  ([#5601](https://github.com/facebook/jest/pull/5601))
+* `[jest-jasmine2]` Adds error throwing and descriptive errors to `it`/ `test`
+  for invalid arguments. `[jest-circus]` Adds error throwing and descriptive
+  errors to `it`/ `test` for invalid arguments
+  ([#5558](https://github.com/facebook/jest/pull/5558))
+* `[jest-matcher-utils]` Add `isNot` option to `matcherHint` function
+  ([#5512](https://github.com/facebook/jest/pull/5512))
+* `[jest-config]` Add `<rootDir>` to runtime files not found error report
+  ([#5693](https://github.com/facebook/jest/pull/5693))
+* `[expect]` Make toThrow matcher pass only if Error object is returned from
+  promises ([#5670](https://github.com/facebook/jest/pull/5670))
+* `[expect]` Add isError to utils
+  ([#5670](https://github.com/facebook/jest/pull/5670))
+* `[expect]` Add inverse matchers (`expect.not.arrayContaining`, etc.,
+  [#5517](https://github.com/facebook/jest/pull/5517))
+* `[expect]` `expect.extend` now also extends asymmetric matchers
+  ([#5503](https://github.com/facebook/jest/pull/5503))
+* `[jest-mock]` Update `spyOnProperty` to support spying on the prototype chain
+  ([#5753](https://github.com/facebook/jest/pull/5753))
+* `[jest-mock]` Add tracking of return values in the `mock` property
+  ([#5752](https://github.com/facebook/jest/pull/5752))
+* `[jest-mock]` Add tracking of thrown errors in the `mock` property
+  ([5764](https://github.com/facebook/jest/pull/5764))
+* `[expect]`Add nthCalledWith spy matcher
+  ([#5605](https://github.com/facebook/jest/pull/5605))
+* `[jest-cli]` Add `isSerial` property that runners can expose to specify that
+  they can not run in parallel
+  ([#5706](https://github.com/facebook/jest/pull/5706))
+* `[jest-cli]` Interactive Snapshot Mode improvements
+  ([#5864](https://github.com/facebook/jest/pull/5864))
 
 ### Fixes
 
+* `[jest-runner]` Assign `process.env.JEST_WORKER_ID="1"` when in runInBand mode
+  ([#5860](https://github.com/facebook/jest/pull/5860))
+* `[jest-cli]` Add descriptive error message when trying to use
+  `globalSetup`/`globalTeardown` file that doesn't export a function.
+  ([#5835](https://github.com/facebook/jest/pull/5835))
+* `[expect]` Do not rely on `instanceof RegExp`, since it will not work for
+  RegExps created inside of a different VM
+  ([#5729](https://github.com/facebook/jest/pull/5729))
+* `[jest-resolve]` Update node module resolution algorithm to correctly handle
+  symlinked paths ([#5085](https://github.com/facebook/jest/pull/5085))
+* `[jest-editor-support]` Update `Settings` to use spawn in shell option
+  ([#5658](https://github.com/facebook/jest/pull/5658))
+* `[jest-cli]` Improve the error message when 2 projects resolve to the same
+  config ([#5674](https://github.com/facebook/jest/pull/5674))
+* `[jest-runtime]` remove retainLines from coverage instrumentation
+  ([#5692](https://github.com/facebook/jest/pull/5692))
+* `[jest-cli]` Fix update snapshot issue when using watchAll
+  ([#5696](https://github.com/facebook/jest/pull/5696))
+* `[expect]` Fix rejects.not matcher
+  ([#5670](https://github.com/facebook/jest/pull/5670))
+* `[jest-runtime]` Prevent Babel warnings on large files
+  ([#5702](https://github.com/facebook/jest/pull/5702))
+* `[jest-mock]` Prevent `mockRejectedValue` from causing unhandled rejection
+  ([#5720](https://github.com/facebook/jest/pull/5720))
+* `[pretty-format]` Handle React fragments better
+  ([#5816](https://github.com/facebook/jest/pull/5816))
+
+### Chore & Maintenance
+
+* `[#5858]` Run Prettier on compiled output
+  ([#5858](https://github.com/facebook/jest/pull/3497))
+* `[#5708]` Add fileChange hook for plugins
+  ([#5708](https://github.com/facebook/jest/pull/5708))
+* `[docs]` Add docs on using `jest.mock(...)`
+  ([#5648](https://github.com/facebook/jest/pull/5648))
+* `[docs]` Mention Jest Puppeteer Preset
+  ([#5722](https://github.com/facebook/jest/pull/5722))
+* `[docs]` Add jest-community section to website
+  ([#5675](https://github.com/facebook/jest/pull/5675))
+* `[docs]` Add versioned docs for v22.4
+  ([##5733](https://github.com/facebook/jest/pull/#5733))
+
+## 22.4.2
+
+### Fixes
+
+* `[jest-haste-map]` Recreate Haste map when deserialization fails
+  ([#5642](https://github.com/facebook/jest/pull/5642))
+
+## 22.4.1
+
+### Fixes
+
+* `[jest-haste-map]` Parallelize Watchman calls in crawler
+  ([#5640](https://github.com/facebook/jest/pull/5640))
+* `[jest-editor-support]` Update TypeScript definitions
+  ([#5625](https://github.com/facebook/jest/pull/5625))
 * `[babel-jest]` Remove `retainLines` argument to babel.
   ([#5594](https://github.com/facebook/jest/pull/5594))
+
+### Features
+
+* `[jest-runtime]` Provide `require.main` property set to module with test suite
+  ([#5618](https://github.com/facebook/jest/pull/5618))
 
 ### Chore & Maintenance
 
@@ -16,6 +110,8 @@
   ([#5622](https://github.com/facebook/jest/pull/5622))
 * `[docs]` Update to use yarn
   ([#5624](https://github.com/facebook/jest/pull/5624))
+* `[docs]` Add how to mock scoped modules to Manual Mocks doc
+  ([#5638](https://github.com/facebook/jest/pull/5638))
 
 ## 22.4.0
 
@@ -53,6 +149,11 @@
   ([#5609](https://github.com/facebook/jest/pull/5609))
 * `[docs]` Add a documentation note for project `displayName` configuration
   ([#5600](https://github.com/facebook/jest/pull/5600))
+
+### Chore & Maintenance
+
+* `[docs]` Update automatic mocks documentation
+  ([#5630](https://github.com/facebook/jest/pull/5630))
 
 ## jest 22.3.0
 
@@ -125,6 +226,8 @@
 * `[diff-sequences]` New package compares items in two sequences to find a
   **longest common subsequence**.
   ([#5407](https://github.com/facebook/jest/pull/5407))
+* `[jest-matcher-utils]` Add `comment` option to `matcherHint` function
+  ([#5437](https://github.com/facebook/jest/pull/5437))
 * `[jest-config]` Allow lastComit and changedFilesWithAncestor via JSON config
   ([#5476](https://github.com/facebook/jest/pull/5476))
 * `[jest-util]` Add deletion to `process.env` as well
@@ -150,8 +253,8 @@
   malformed source-map. ([#5405](https://github.com/facebook/jest/pull/5405)).
 * `[jest]` Add `import-local` to `jest` package.
   ([#5353](https://github.com/facebook/jest/pull/5353))
-* `[expect]` Support class instances in `.toHaveProperty()` matcher.
-  ([#5367](https://github.com/facebook/jest/pull/5367))
+* `[expect]` Support class instances in `.toHaveProperty()` and `.toMatchObject`
+  matcher. ([#5367](https://github.com/facebook/jest/pull/5367))
 * `[jest-cli]` Fix npm update command for snapshot summary.
   ([#5376](https://github.com/facebook/jest/pull/5376),
   [5389](https://github.com/facebook/jest/pull/5389/))